FIxed the issue no 18
FIxed the issue no 18 and updated the code
