+++ DemoSurveysAndroid/.idea/gradle.xml
@@ -9,7 +9,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/opgsdk_lite_v1.5" />
+            <option value="$PROJECT_DIR$/opgsdk_lite_v1.6" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
