+++ DemoSurveysAndroid/.idea/modules.xml
@@ -4,10 +4,13 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/DemoSurveysAndroid.iml" filepath="$PROJECT_DIR$/DemoSurveysAndroid.iml" />
       <module fileurl="file://$PROJECT_DIR$/DemoSurveysAndroid.iml" filepath="$PROJECT_DIR$/DemoSurveysAndroid.iml" />
+      <module fileurl="file://$PROJECT_DIR$/DemoSurveysAndroid.iml" filepath="$PROJECT_DIR$/DemoSurveysAndroid.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/opgsdk_lite_v1.5/opgsdk_lite_v1.5.iml" filepath="$PROJECT_DIR$/opgsdk_lite_v1.5/opgsdk_lite_v1.5.iml" />
       <module fileurl="file://$PROJECT_DIR$/opgsdk_lite_v1.5/opgsdk_lite_v1.5.iml" filepath="$PROJECT_DIR$/opgsdk_lite_v1.5/opgsdk_lite_v1.5.iml" />
+      <module fileurl="file://$PROJECT_DIR$/opgsdk_lite_v1.6/opgsdk_lite_v1.6.iml" filepath="$PROJECT_DIR$/opgsdk_lite_v1.6/opgsdk_lite_v1.6.iml" />
     </modules>
   </component>
 </project>
