+++ DemoSurveysAndroid/app/build.gradle
@@ -30,7 +30,6 @@ android {
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
     // testCompile 'junit:junit:4.12'
-    compile project(':opgsdk_lite_v1.5')
     compile 'log4j:log4j:1.2.17'
     compile 'com.android.support:appcompat-v7:23.2.1'
     compile 'com.android.support:support-v4:23.2.1'
@@ -39,5 +38,6 @@ dependencies {
     compile 'com.google.firebase:firebase-messaging:9.8.0'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.google.code.gson:gson:2.8.0'
+    compile project(':opgsdk_lite_v1.6')
 }
 apply plugin: 'com.google.gms.google-services'
