+++ DemoSurveysAndroid/app/src/main/res/layout/activity_get_panelist_panel.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.onepointglobal.mysurveysn.GetPanelistPanel">
+    tools:context="com.onepointglobal.mysurveysn.GetPanellistPanel">
     <TextView
         android:id="@+id/panelist_panel_output"
         android:layout_width="match_parent"
