--- DemoSurveysAndroid/opgsdk_lite_v1.5/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('opgsdk_lite_v1.5.aar'))
