+++ DemoSurveysAndroid/opgsdk_lite_v1.6/build.gradle
@@ -1,0 +1,2 @@
+configurations.maybeCreate("default")
+artifacts.add("default", file('opgsdk_lite_v1.6.aar'))
