+++ DemoSurveysAndroid/settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':opgsdk_lite_v1.5'
+include ':app', ':opgsdk_lite_v1.6'
